🐙ueli🎆26🌙06☀️04🚩313
🎆26🌙07☀️19⏰13⌚26⏱️05
🗺️Plan Norm absolute completeness to close remaining Eurocode and energy-family gate gaps
🌡️Scaffold mathematical entropy crate with FFT, discrete estimators, and numeric infrastructure
🕸️Add framework OS StudioRunner for headless media-graph execution via os run
📐Expand EN 1993–1999 norm crates with real clause physics toward absolute completeness
🎲Grow mathematical sampling engine and CAS/number/polynomial modules for NetworkX parity
📚Add storybook styling and puzzle 3d/5d/infinite Playwright coverage specs
🧪Wire leveled test targets across ui, cad, kernel, compose, and vcs packages
🛠️Register root script.ts os run, trim .devcontainer ports, and fix repo vscode vite config

📊uloc➕12841✏️1325➖1🟰14167
🧾412k➕650✏️7🟰657
🦀344k➕12043✏️1240🟰13283
🟦164k➕101✏️76🟰177
🔵95k
🟣32k
🐍31k
🎨21k
🟨16k
📝12k
⚙️6k➕47✏️2➖1🟰50
🛢️3k
🌐3k
📋1k
💠893
🐚586
📑57
🐳43

+++ mathematical/number/rs/src/natural.rs
@@ -441,10 +441,10 @@ impl Natural {
             x = next;
         }
         // Correct rounding: Newton converges to floor(root) or floor(root)+1; step down until x^n <= self.
-        while x.pow(n) > *self {
+        while x.pow(n as u64) > *self {
             x = x.checked_sub(&Self::one()).unwrap_or_else(Self::zero);
         }
-        while x.add(&Self::one()).pow(n) <= *self {
+        while x.add(&Self::one()).pow(n as u64) <= *self {
             x = x.add(&Self::one());
         }
         x
